make node tests pass
